requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pygame-ce==2.4.0
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
twitchAPI==4.2.0
typing_extensions==4.9.0
yarl==1.9.4
